Created a utility function for lauching a web browser with a specified url.
