+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/META-INF/MANIFEST.MF
@@ -28,6 +28,7 @@ Import-Package: com.google.api.client.auth.oauth2,
  com.google.cloud.tools.eclipse.login,
  com.google.cloud.tools.eclipse.sdk,
  com.google.cloud.tools.eclipse.sdk.ui,
+ com.google.cloud.tools.eclipse.ui.util,
  org.eclipse.core.runtime;bundle-symbolic-name:="org.eclipse.core.runtime",
  org.eclipse.core.runtime.preferences;version="3.3.0",
  org.osgi.framework;version="1.8.0",
