+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/messages.properties
@@ -15,4 +15,4 @@ deploy.failed.error.message=Deploy failed.
 cloudsdk.process.failed=Process exited with error code {0}
 save.credential.failed=Error temporarily saving credential
 brower.launch.failed=Error launching deployed app in browser
-invalid.url=Invalid URL
+browser.name=App Engine Deploy - {0}
