+++ plugins/com.google.cloud.tools.eclipse.preferences/META-INF/MANIFEST.MF
@@ -12,4 +12,5 @@ Bundle-ActivationPolicy: lazy
 Bundle-Localization: plugin
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime
-Import-Package: org.eclipse.core.resources
+Import-Package: com.google.cloud.tools.eclipse.ui.util,
+ org.eclipse.core.resources
