+++ plugins/com.google.cloud.tools.eclipse.sdk.ui/META-INF/MANIFEST.MF
@@ -14,7 +14,8 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.107.0",
  org.eclipse.e4.core.contexts;bundle-version="1.4.0",
  com.google.cloud.tools.eclipse.sdk;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.preferences
-Import-Package: org.eclipse.osgi.util;version="1.1.0",
+Import-Package: com.google.cloud.tools.eclipse.ui.util,
+ org.eclipse.osgi.util;version="1.1.0",
  org.eclipse.ui.console,
  org.osgi.framework;version="1.8.0"
 Export-Package: com.google.cloud.tools.eclipse.sdk.ui,
