+++ plugins/com.google.cloud.tools.eclipse.sdk.ui/src/com/google/cloud/tools/eclipse/sdk/ui/preferences/messages.properties
@@ -1,6 +1,4 @@
 CloudSdkRequired=Google Cloud Tools for Eclipse requires the <a href="https://cloud.google.com/sdk/">Google Cloud SDK</a> with the <a href="https://cloud.google.com/sdk/docs/managing-components">App Engine Java components</a>.
-CloudSdkPreferencePage_3=URL malformed
-CloudSdkPreferencePage_4=Cannot launch a browser
 SdkLocation=&SDK location:
 CloudSdkNotFound=Cloud SDK not found in {0}
 CloudSdkOutOfDate=Installed Cloud SDK is too old. Run `gcloud components update` and try again.
