+++ plugins/com.google.cloud.tools.eclipse.ui.util/META-INF/MANIFEST.MF
@@ -21,6 +21,7 @@ Import-Package: com.google.cloud.tools.eclipse.ui.util,
  org.eclipse.core.expressions,
  org.eclipse.core.resources,
  org.eclipse.core.runtime;bundle-symbolic-name:="org.eclipse.core.runtime",
+ org.eclipse.core.runtime.jobs,
  org.eclipse.jface.action,
  org.eclipse.jface.dialogs,
  org.eclipse.jface.resource,
