+++ plugins/com.google.cloud.tools.eclipse.ui.util/src/com/google/cloud/tools/eclipse/ui/util/messages.properties
@@ -2,6 +2,7 @@ bucket.name.invalid=Invalid bucket name
 bucket.name.invalid.long=Bucket name must be between 3 and 63 characters. It can have lowercase \
 letters, numbers, hyphens (-), underscores (_) . Dot (.) can be used to form a valid top-level \
 domain (e.g., example.com). It must start and end with a number or a letter.
+invalid.url=Invalid URL
 project.id.invalid=Invalid project ID
 project.id.empty=Empty project ID
 project.id.invalid.long=Project ID must be between 6 and 30 characters. It can have lowercase \
